Move SimplestStorageProvider and associated type implementations into own modules
