+++ p2panda-rs/src/test_utils/db/stores/entry.rs
@@ -9,8 +9,8 @@ use crate::identity::Author;
 use crate::operation::AsOperation;
 use crate::schema::SchemaId;
 use crate::storage_provider::errors::EntryStorageError;
-use crate::storage_provider::traits::test_utils::{SimplestStorageProvider, StorageEntry};
 use crate::storage_provider::traits::{AsStorageEntry, EntryStore};
+use crate::test_utils::db::{SimplestStorageProvider, StorageEntry};
 
 /// Implement `EntryStore` trait on `SimplestStorageProvider`
 #[async_trait]
@@ -142,10 +142,9 @@ pub mod tests {
     use crate::identity::{Author, KeyPair};
     use crate::operation::OperationEncoded;
     use crate::schema::SchemaId;
-    use crate::storage_provider::traits::test_utils::{
-        test_db, SimplestStorageProvider, StorageEntry, TestStore,
-    };
+    use crate::storage_provider::traits::test_utils::{test_db, TestStore};
     use crate::storage_provider::traits::{AsStorageEntry, EntryStore};
+    use crate::test_utils::db::{SimplestStorageProvider, StorageEntry};
     use crate::test_utils::fixtures::{
         entry, entry_signed_encoded, key_pair, operation_encoded, random_key_pair, schema,
     };
