+++ p2panda-rs/src/test_utils/db/types/entry.rs
@@ -1,0 +1,92 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+
+use crate::entry::{decode_entry, Entry, EntrySigned, LogId, SeqNum};
+use crate::hash::Hash;
+use crate::identity::Author;
+use crate::operation::{Operation, OperationEncoded};
+use crate::storage_provider::errors::EntryStorageError;
+use crate::storage_provider::traits::AsStorageEntry;
+use crate::storage_provider::ValidationError;
+use crate::Validate;
+
+/// A struct which represents an entry and operation pair in storage as a concatenated string.
+#[derive(Debug, Clone, PartialEq)]
+pub struct StorageEntry(String);
+
+impl StorageEntry {
+    pub fn entry_decoded(&self) -> Entry {
+        // Unwrapping as validation occurs in constructor.
+        decode_entry(&self.entry_signed(), self.operation_encoded().as_ref()).unwrap()
+    }
+
+    pub fn entry_signed(&self) -> EntrySigned {
+        let params: Vec<&str> = self.0.split('-').collect();
+        EntrySigned::new(params[0]).unwrap()
+    }
+
+    pub fn operation_encoded(&self) -> Option<OperationEncoded> {
+        let params: Vec<&str> = self.0.split('-').collect();
+        Some(OperationEncoded::new(params[1]).unwrap())
+    }
+}
+
+/// Implement `AsStorageEntry` trait for `StorageEntry`
+impl AsStorageEntry for StorageEntry {
+    type AsStorageEntryError = EntryStorageError;
+
+    fn new(
+        entry: &EntrySigned,
+        operation: &OperationEncoded,
+    ) -> Result<Self, Self::AsStorageEntryError> {
+        let entry_string = format!("{}-{}", entry.as_str(), operation.as_str());
+        let storage_entry = Self(entry_string);
+        storage_entry.validate()?;
+        Ok(storage_entry)
+    }
+
+    fn author(&self) -> Author {
+        self.entry_signed().author()
+    }
+
+    fn hash(&self) -> Hash {
+        self.entry_signed().hash()
+    }
+
+    fn entry_bytes(&self) -> Vec<u8> {
+        self.entry_signed().to_bytes()
+    }
+
+    fn backlink_hash(&self) -> Option<Hash> {
+        self.entry_decoded().backlink_hash().cloned()
+    }
+
+    fn skiplink_hash(&self) -> Option<Hash> {
+        self.entry_decoded().skiplink_hash().cloned()
+    }
+
+    fn seq_num(&self) -> SeqNum {
+        *self.entry_decoded().seq_num()
+    }
+
+    fn log_id(&self) -> LogId {
+        *self.entry_decoded().log_id()
+    }
+
+    fn operation(&self) -> Operation {
+        let operation_encoded = self.operation_encoded().unwrap();
+        Operation::from(&operation_encoded)
+    }
+}
+
+impl Validate for StorageEntry {
+    type Error = ValidationError;
+
+    fn validate(&self) -> Result<(), Self::Error> {
+        self.entry_signed().validate()?;
+        if let Some(operation) = self.operation_encoded() {
+            operation.validate()?;
+        }
+        decode_entry(&self.entry_signed(), self.operation_encoded().as_ref())?;
+        Ok(())
+    }
+}
