+++ p2panda-rs/src/test_utils/db/types/response.rs
@@ -1,0 +1,57 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+
+use crate::entry::{LogId, SeqNum};
+use crate::hash::Hash;
+use crate::storage_provider::traits::{AsEntryArgsResponse, AsPublishEntryResponse};
+
+#[derive(Debug, Clone, PartialEq)]
+pub struct PublishEntryResponse {
+    pub entry_hash_backlink: Option<Hash>,
+    pub entry_hash_skiplink: Option<Hash>,
+    pub seq_num: SeqNum,
+    pub log_id: LogId,
+}
+
+impl AsPublishEntryResponse for PublishEntryResponse {
+    /// Just the constructor method is defined here as all we need this trait for
+    /// is constructing entry args to be returned from the default trait methods.
+    fn new(
+        entry_hash_backlink: Option<Hash>,
+        entry_hash_skiplink: Option<Hash>,
+        seq_num: SeqNum,
+        log_id: LogId,
+    ) -> Self {
+        Self {
+            entry_hash_backlink,
+            entry_hash_skiplink,
+            seq_num,
+            log_id,
+        }
+    }
+}
+
+#[derive(Debug, Clone, PartialEq)]
+pub struct EntryArgsResponse {
+    pub entry_hash_backlink: Option<Hash>,
+    pub entry_hash_skiplink: Option<Hash>,
+    pub seq_num: SeqNum,
+    pub log_id: LogId,
+}
+
+impl AsEntryArgsResponse for EntryArgsResponse {
+    /// Just the constructor method is defined here as all we need this trait for
+    /// is constructing entry args to be returned from the default trait methods.
+    fn new(
+        entry_hash_backlink: Option<Hash>,
+        entry_hash_skiplink: Option<Hash>,
+        seq_num: SeqNum,
+        log_id: LogId,
+    ) -> Self {
+        Self {
+            entry_hash_backlink,
+            entry_hash_skiplink,
+            seq_num,
+            log_id,
+        }
+    }
+}
